adding AdoptionHelper methods
